requirements.txt
asgiref==3.5.2
attrs==22.1.0
backports.zoneinfo==0.2.1
Django==4.1.1
django-cors-headers==3.13.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
drf-spectacular==0.23.1
importlib-resources==5.9.0
inflection==0.5.1
jsonschema==4.16.0
Pillow==9.2.0
pkgutil_resolve_name==1.3.10
PyJWT==2.4.0
pyrsistent==0.18.1
python-decouple==3.6
pytz==2022.2.1
PyYAML==6.0
sqlparse==0.4.2
uritemplate==4.1.1
zipp==3.8.1
